PRINT_EXIT_VALUE: Honour the option in bodies of anonymous functions
Before this patch, only the anonymous function's return value would be subject
to the option; with this patch, each individual sublist within the function is
subject to the option, as well as the anonymous function itself.

That does lead to duplicate output when the last command in an anonymous
function returns non-zero.